Add Fn Lock functionality with fn+space
TG(1) would normally do this but I wanted the fn lock to live on the
fucntion layer (otherwise it's a waste of a key).  The problem with this
is that calling TG(1) from within layer 1 gets reset when the MO(1) key
is released.  Therefore, We effectively have to recreate the
functionality with an exception for the fn lock toggle.